+++ src/main/java/duke/task/ToDo.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.task;
 
 public class ToDo extends Task {
     public ToDo(String isDone, String description) {
